Conflicts:
	overlay/d3d10.cpp
	overlay/overlay.h
	overlay/overlay.pro
	overlay_gl/overlay.c

+++ src/mumble/Overlay.cpp
@@ -32,6 +32,7 @@
 #include "OverlayText.h"
 #include "User.h"
 #include "Channel.h"
+#include "ClientUser.h"
 #include "Global.h"
 #include "Message.h"
 #include "Database.h"
@@ -199,7 +200,7 @@ void Overlay::verifyTexture(ClientUser *cp, bool allowupdate) {
 	if (! cp->qbaTexture.isEmpty()) {
 		bool valid = true;
 
-		if (cp->qbaTexture.length() < sizeof(unsigned int)) {
+		if (cp->qbaTexture.length() < static_cast<int>(sizeof(unsigned int))) {
 			valid = false;
 		} else if (qFromBigEndian<unsigned int>(reinterpret_cast<const unsigned char *>(cp->qbaTexture.constData())) == 600 * 60 * 4) {
 			QByteArray qba = qUncompress(cp->qbaTexture);
